Changed More Stuff

+++ RRDv2/Models/Floor.cs
@@ -9,7 +9,7 @@ namespace RRDv2.Models
     {
         public int Id { get; set; }
         public int HotelId { get; set; }
-        public ushort FloorNum { get; set; }
+        public int FloorNum { get; set; }
         public virtual Hotel Hotel { get; set; }
         public virtual ICollection<Room> Rooms { get; set; }
     }
